# isme-nest-serrve.swagger.json
{
	"info": {
		"title": "isme-nest-serrve",
		"description": "",
		"version": "1.0.0"
	},
...
	"paths": {
...
		"/user/detail": {
			"get": {
				"summary": "用户详情",
				"deprecated": false,
				"description": "",
				"tags": ["用户"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "object",
									"properties": {
										"id": {
											"type": "integer"
										},
										"username": {
											"type": "string"
										},
										"enable": {
											"type": "boolean"
										},
										"createTime": {
											"type": "string"
										},
										"updateTime": {
											"type": "string"
										},
										"profile": {
											"type": "object",
											"properties": {
												"id": {
													"type": "integer"
												},
												"nickName": {
													"type": "string"
												},
												"gender": {
													"type": "null"
												},
												"avatar": {
													"type": "string"
												},
												"address": {
													"type": "null"
												},
												"email": {
													"type": "null"
												},
												"userId": {
													"type": "integer"
												}
											},
											"required": ["id", "nickName", "gender", "avatar", "address", "email",
												"userId"
											]
										},
										"roles": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "integer"
													},
													"code": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"enable": {
														"type": "boolean"
													}
												},
												"required": ["id", "code", "name", "enable"]
											}
										},
										"currentRole": {
											"type": "object",
											"properties": {
												"id": {
													"type": "integer"
												},
												"code": {
													"type": "string"
												},
												"name": {
													"type": "string"
												},
												"enable": {
													"type": "boolean"
												}
											},
											"required": ["id", "code", "name", "enable"]
										}
									},
									"required": ["id", "username", "enable", "createTime", "updateTime", "profile",
										"roles", "currentRole"
									]
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/user": {
			"get": {
				"summary": "用户列表-分页",
				"deprecated": false,
				"description": "",
				"tags": ["用户"],
				"parameters": [{
					"name": "username",
					"in": "query",
					"description": "",
					"required": false,
					"type": "string",
					"x-example": "admin"
				}, {
					"name": "pageNo",
					"in": "query",
					"description": "",
					"required": true,
					"type": "integer",
					"x-example": 1
				}, {
					"name": "pageSize",
					"in": "query",
					"description": "",
					"required": true,
					"type": "integer",
					"x-example": 10
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "object",
									"properties": {
										"pageData": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "integer"
													},
													"username": {
														"type": "string"
													},
													"enable": {
														"type": "boolean"
													},
													"createTime": {
														"type": "string"
													},
													"updateTime": {
														"type": "string"
													},
													"roles": {
														"type": "array",
														"items": {
															"type": "object",
															"properties": {
																"id": {
																	"type": "integer"
																},
																"code": {
																	"type": "string"
																},
																"name": {
																	"type": "string"
																},
																"enable": {
																	"type": "boolean"
																}
															},
															"required": ["id", "code", "name", "enable"]
														}
													},
													"gender": {
														"type": "null"
													},
													"avatar": {
														"type": "string"
													},
													"address": {
														"type": "null"
													},
													"email": {
														"type": "null"
													}
												}
											}
										},
										"total": {
											"type": "integer"
										}
									},
									"required": ["pageData", "total"]
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			},
			"post": {
				"summary": "新增用户",
				"deprecated": false,
				"description": "",
				"tags": ["用户"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/user/{id}": {
			"delete": {
				"summary": "删除用户",
				"deprecated": false,
				"description": "",
				"tags": ["用户"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			},
			"patch": {
				"summary": "修改用户",
				"deprecated": false,
				"description": "",
				"tags": ["用户"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/user/password/reset/{id}": {
			"patch": {
				"summary": "重置用户密码",
				"deprecated": false,
				"description": "",
				"tags": ["用户"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
...
	},
	"swagger": "2.0",
	"basePath": null,
...
	"securityDefinitions": {
		"bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
...
}